DCON-4889 Extract shared postRequestCleanup middleware, dedupe 3 copies
mcpRequestCleanup (app.js) was a third verbatim copy of the res.once
('finish', ...) audit-flush logic already duplicated between
graphqlPostRequestCleanup and graphqlV2PostRequestCleanup -- same
container/postRequestProcessor/requestSpecificCache lookup and
executeAsync/clearAsync calls, only cosmetic naming differed. Extracted
to src/middleware/postRequestCleanup.middleware.js (factory, matching
this repo's existing middleware/*.middleware.js convention) so a future
fix to this flush logic only needs to be applied once.
